Split dispatcher tests into helper + dispatcher files; mirror BuilderTests pattern
